+++ aiobungie/error.py
@@ -59,7 +59,7 @@ class HTTPException(AiobungieError):
     """Exception for handling `aiobungie.rest.RESTClient` requests errors."""
 
     message: str = attr.field(default="")
-    long_message: typing.Optional[str] = attr.field(default=None)
+    long_message: str = attr.field(default="")
 
 
 @attr.define(auto_exc=True, repr=False, weakref_slot=False, kw_only=True)
@@ -71,9 +71,12 @@ class RateLimitedError(HTTPException):
     url: str = attr.field(default="")
 
 
-@attr.define(auto_exc=True, repr=False, weakref_slot=False, kw_only=True)
-class InternalServerError(AiobungieError):
+@attr.define(auto_exc=True, repr=False, weakref_slot=False)
+class InternalServerError(HTTPException):
     """Raised for other 5xx errors."""
+
+    message: str = attr.field(default="")
+    long_message: str = attr.field(default="")
 
 
 class NotFound(AiobungieError):